+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.1.19 (19-11-2020)
 - Added 'Strip Whitespace From Export Paths' option to settings (enabled by default).
+- Added the FileDataID field to ModelPlacementInformation CSV files.
 
 0.1.18 (10-11-2020)
 - Added 'Texture Alpha' option to model exporter, allowing binary control of texture transparency on models.